requirements.txt
Django==3.2.12
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0